libjson/json_util.c: add missing cast. CID 1404169

+++ gdal/ogr/ogrsf_frmts/geojson/libjson/json_util.c
@@ -293,7 +293,7 @@ const char *json_type_to_name(enum json_type o_type)
 	int o_type_int = (int)o_type;
 	if (o_type_int < 0 || o_type_int >= (int)NELEM(json_type_name))
 	{
-		MC_ERROR("json_type_to_name: type %d is out of range [0,%d]\n", o_type, NELEM(json_type_name));
+		MC_ERROR("json_type_to_name: type %d is out of range [0,%d]\n", o_type, (int)NELEM(json_type_name));
 		return NULL;
 	}
 	return json_type_name[o_type];
